fix: moved login api middleware folder out of frontend
